deps/v8/test/inspector/debugger/terminate-execution-on-pause-expected.txt

Tests Runtime.terminateExecution on pause

Running test: testTerminateOnDebugger

Running test: testTerminateAtBreakpoint
